requirements.txt
torch>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
tqdm>=4.65.0
tensorboard>=2.13.0
gymnasium>=0.28.0
